cv/glasses/requirements.txt
onvif2-zeep
